Improve tests
The ambition with this commit is to transform the messy tests
into code that is at least not painful to read.

From here on we can continue to carefully go through what is
being tested for, and to think about what we should really test.

The .out file has been carefully inspected manually line-by-line
and the output is expected.

Changes:

* Use lower_case consistently
* Harmonize naming
* Remove redundant tests
* Remove unused extra columns
  Such columns just make the tests harder to understand,
  without adding any complexity or any other value.
* Fix a lot of broken comments
  Such comments have probably been correct at some point,
  but someone has forgot to update them after changing the code.
* Use multi-line syntax for larger CREATE TABLE statements
  This makes the column names stand out better.
  Keep single line CREATE TABLE for simple cases.
* Remove "-- ok" and "-- fail" at end of lines
  Instead, group "ok" and "fail" cases,
  to avoid needing a comment on each case.