PG-1002: WAL encryption requires explicit configuration (#36)
Instead of automatically creating a default keyring, from now on
we require users to expicitly create a WAL key. Most of these
steps were required even without change anyway, as the default
configuration was highly unsecore.

This eliminates the possiblity of users forgetting to change the
unsecure default, ending up with an encryption that doesn't work
in practice.

The required steps are outlined in the new tap test, that tries
to enable wal encryption:

* Enable the extension in at least one database
* Create a global key provider
* Create a global principal key
* Create the WAL key using the new `pg_tde_create_wal_key()` function
* Set `pg_tde.wal_encrypt = 1` in the conf file or with `ALTER SYSTEM`
* Restart the server

Setting the GUC variable to ON without the previous steps results
in the startup failing with an error message explaining the requirements.