PG-1256: Default principal key functionality
This commit adds a new function, pg_tde_set_default_principal_key which
can be used to mark a global principal key as the default principal key.

If a default principal key is set, databases that do not have a
principal key assigned automatically get assigned the default key during
the creation of the first encrypted database object.

When the default key is changed (by a repeated call to the same function
with a different key name), all databases that currently use the default
principal key move to the new default principal key. This is also true
for databases that set the use of this key manually.

Default keys are independent of pg_tde.inherit_global_providers, and
can even be used when that setting is turned off. Setting up a default
principal key requires global key admin permissions, but after that
any database will be able to use encryption with default tables without
any specific permissions required.

Removal of the default key is not implemented: that will be part of the
principal key deletion functionality, PG-1257.

Internally the default key is implemented as a special database, which
uses the OID of the default tablespace. When the default key is set,
pg_tde internally assigns this principal key to the default oid.

When a database has no principal key configured, pg_tde checks if the
default oid has a principal key configured, and also configures the
same principal key to the current database if yes.