Simplify logic for stopping the server in configure-tde-server.sh
I run multiple instances of PostgreSQL on my machine so I would rather
not have it rely on using pgrep to detect if the server is running.

To make this new code more reliable we add set -e so the script aborts
directly e.g. if the port is already in use.