PG-1373 Use memcpy() and strlcpy() over strncpy()
Since strncpy() does not zero terminate strings when it stops at the
end of the buffer it is a very dangerous function to use. And
instead of manually do the zero termination after calling strncpy()
we use safer APIs like strlen() + memcpy() and strlcpy() and in
a couple of places just a plain memcpy() where both sides have a
known length and we want to make sure the zero padding is correct.

We also need to call memset() or palloc0() where we want the whole
struct initialized because we plan to write it to disk later.

A couple of the strncpy() were broken while others were safe but
we fix all cases to be certain.