PG-1222 Do not write keys for temporary relations to disk
Since temporary relations in PostgreSQL are per backend we do not have
to write them to disk. This is an optimization but also a correctness
fix since the RelFileNumber of at temporary relations are only unqiue
per tablespace.

A downside of never writing the keys to disk is that we cannot see
if other backends have encrypted relations or not. But comapred
to the alternative of adding a backend field per the keys in the
files and implementing proper cleanup this seems like a lesser evil,
especially since keys are right now also written to WAL for
non-temporary relations.