PG-1446 Add missing lock to pg_tde_create_wal_key()
Make sure that we lock the key file before adding a WAL key. It should
be harmless since this happens very early in the start but if we ever
add the ability to rotate WAL keys while running this will become an
issue.

Additionally we add asserts to make sure we hold the right lock level
when we open the key map file.