Use restart TAP helper and do not assert result
There is a restart function so there is not and need to call first stop
and then start. And since by default a start, stop or restart call does
not return on error it is totally pointless to assert anything about the
return value. And since PostgreSQL's own tests also are fine with just
bailing out on error we do the same.

While at it we also always call these three functions without
parentheses to be consistent.