PG-1507 PG-1508 Disallow ALTER TABLE when there is a mix of encrypted and unencrypted
Since ALTER TABLE can modify multiple due to inheritance or typed tables
which can for example result in TOAST tables being created for some or all
of the modified tables while the event trigger is only fired once we cannot
rely on the event stack to make sure we get the correct encryption status.

Our solution is to be cautious and only modify tables when all tables
with storage are either encrypt or not encrypted. If there is a mix we
will throw an error. The result of this is also used to properly inform
the SMGR of the current encryption status. We apply the same logic to
ALTER TYPE for typed tables.

A possibility for future improvement would be to limit this check only
to commands which recurse down to child tables and/or can create new
relfilenodes.