Improve function names and signatures for encryption
The naming was quite inconsistent and, in the case of ECB vs CTR,
misleading. So let's make it more consistent plus remove the legacy
macros for the WAL encryption which were only used to slightly improve
debug logging.

Also do not supply any IV to ECB since it does not use any IV.