PG-1651 Do not try the fetch key from old provider when changing provider settings
We used to verify that all keys are still accessible by fetching the key
from cache and if not cached from the current provider settings when
changing the key provider setting but this makes things fragile and
prone to user error in the case where a principal key is not cached.

Instead we can rely on the AEAD tag in the header of each file, which
has two advantages 1) the cached and non-cache code path works the same
and 2) we do not need to care about if the old provider is accessible,
only that the new one is and that it contains the right key (i.e. one
that can verify the header tag).