Remove provider_ prefix from parameter names
There was no reason for name and type to be prefixed in the add/change
key provider functions while options was not.

The output from the list_providers functions had them for name and type,
but not id nor options. So remove the prefixes there aswell.

In other functions the prefixes serve to show what parameters is about
the provider and which are about something else.