PG-1842 Minor optimization and cleanup of redo of SMGR create
The code we ran when redoing a smgrcreate() was overly complex and not
necessary to run. If the file descriptor was open we skipped it but if
not we ran a bit of pointless code since the key creation is handled by
it is own WAL record which is always before the SMGR creation.

Also improve some outdated comments.