PG-1870 Pre-generate pg_tde files for TAP tests
This is currently only added for meson, but could also be added for
make.

This has to be setup before the actual TAP tests are run as they are run
in parallel and as such would all try to setup the template at the same
time if we let them use the same folder for it without it being
pre-generated.

This seems to shorten the test suite run-time by ~25% on my laptop, so
it seems worth doing.