Respect extra arguments to initdb in TDE_MODE
Previously we ignored the extra arguments to initdb when initializing
the pg_tde directory and just copied the directory from a database
initialized without the extra arguments, or if available from the cache.

Now make sure that when extra arguments are supplied that we do not
use the cache and that we copy the pg_tde directory from database
initialized with the extra arguments.

As far as I know this is only relevant to the --allow-group-access flag
but we may as well make the solution generic.